pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "py-mal"
version = "0.1.0"
description = "Examples of prexisting-malware types created in modern Python; purely for education purposes only"
readme = "README.md"
requires-python = ">=3.14"
license = "Unlicense"

dependencies = [
    "click>=8.3.1",
    "mss>=10.1.0",
    "pillow>=12.1.1",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pygame>=2.6.1",
]

[project.scripts]
rev-shell = "rev_shell.cli:main"
rat = "rat.cli:main"

[dependency-groups]
dev = [
  "ruff==0.15.7",
  "pytest>=8.0.0",
  "pytest-mock>=3.15.1",
  "pytest-cov>=7.1.0",
  "pytest-timeout>=2.4.0",
]


[tool.hatch.build.targets.wheel]
packages = ["src/rat", "src/rev_shell"]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"
markers = [
    "smoke: Quick hardware/dependency checks",
    "unit: Isolated logic tests",
    "integration: Socket and protocol tests",
]

[project.urls]
Repository = "https://github.com/sethbarrett50/py_mal"
Issues = "https://github.com/sethbarrett50/py_mal/issues"
